Rework Wire Protocol (#1906)

+++ src/CLR/Include/WireProtocol_HAL_Interface.h
@@ -9,8 +9,24 @@
 #include "WireProtocol.h"
 
 //////////////////////////////////////////
-int WP_ReceiveBytes(uint8_t *ptr, unsigned short *size);
-int WP_TransmitMessage(WP_Message *message);
+
+///
+/// @brief Receives n bytes from the Wire Protocol channel.
+///
+/// @param ptr Pointer to the buffer that will hold the received bytes.
+/// @param size Number of bytes to read. On return it will have the number of bytes actually received.
+/// @return bool true if any bytes where received, false otherwise.
+///
+uint8_t WP_ReceiveBytes(uint8_t *ptr, uint32_t *size);
+
+///
+/// @brief Sends a message through the Wire Protocol channel.
+///
+/// @param message Message to send
+/// @return bool true for transmition succesfull, false otherwise.
+///
+uint8_t WP_TransmitMessage(WP_Message *message);
+
 void WP_CheckAvailableIncomingData();
 
 #endif // _WIREPROTOCOL_HAL_INTERFACE_H_
